requirements.txt
django==2.2.8
gunicorn
django-heroku
requests
djangorestframework
django-cors-headers
python-dotenv==0.14.0
django-extensions
django_postgres_extensions
six